Update jquery.storelocator.js
Corrected 'lengthUnit' option default value from 'm' (not recognized) to 'km'.

Add some checks to disable max distance if the 'maxDistance' variable is null, undefined or 0.

Created a new function called "start" and call it at the initialization process.

When the address input is blank and the form is submitted the start function is called to reinitialize the map. The alert box displaying "The input box was blank" as been removed.

Replaced tab indentations by spaces.