Development commit 1
This is work in progress and most of it probably isn’t working (not
ready for use). I’m switching between computers for development and
figured it would be easier to just create a development branch than
trying to sync files. People will also be able to see changes in
progress. At this point you’ll see I’ve restructured all the files, set
the project up with grunt, prefixed the css, and other changes in the
main JS file.