Easy Debug, sessionStorage and process on field value
Added 2 new settings. The first is a debug setting that defaults to
false, if a user sets it to true, debug information will be rendered to
the console allowing them to know how much processing is actually
happening which is helpful with debugging and getting familiarized with
the plugin. The second setting is to use sessionStorage for geolocation
lookup, for desktops this is good as the user is stationary and if they
are browsing on and off the page, coming back to the page should not
always require 3+ seconds just for that feature to run its process. In
the _start function added some additional code to allow auto processing
if the address field has a value and stop auto geo if that or a query
string param fires.